new features

+++ back/src/main/java/com/parkify/back/repository/MessageRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MessageRepository extends JpaRepository<Message,Integer> {
+public interface MessageRepository extends JpaRepository<Message,Long> {
     List<Message> findByReceiver(User receiver);
 }
