+++ back/src/main/java/com/parkify/back/repository/UserRepository.java
@@ -4,11 +4,11 @@ import com.parkify.back.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     boolean existsByEmail(String email);
     User findByEmail(String email);
     User findByRole(Role role);
-    User findMeById(int id);
+    User findMeById(long id);
 
 
 }
